requirements.txt
alpaca-py
plotly
requests
asyncio
pandas
numpy
matplotlib
TA-lib